4/27/22 - Finalizing for Vizagogo
-Made dialog boxes move with their size
-Made item slots on dialog boxes accurately placed
-Tweaked visual effects on glass on bar
-made glass on bar more accurately move towards customer
-Made difficulty affect order size
-Added a lower difficulty level before 1 for balancing
-Fixed height limit visually for flag
-Added image in game over to represent how far the player got in their playthrough
-Added notification alert telling player to hurry up if they're low on time
-Probably a lot more but its too much to recall everything

+++ Assets/Scripts/Monster.cs
@@ -24,7 +24,7 @@ public class Monster : MonoBehaviour
 
     private float timeWaiting;
 
-    private int irritationFactor;
+    private float irritationFactor;
     public MonsterDrink monsterDrink;
     private bool movingAway;
     private float timer;
@@ -143,12 +143,12 @@ public class Monster : MonoBehaviour
         
     }
 
-    public int getIrritationFactor()
+    public float getIrritationFactor()
     {
         return irritationFactor;
     }
 
-    public void setIrritationFactor(int factor)
+    public void setIrritationFactor(float factor)
     {
         irritationFactor = factor;
     }
